+++ analyzer/src/main/code-gen/UserAgents/ISOLanguageCode/Generate.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # Yet Another UserAgent Analyzer
-# Copyright (C) 2013-2023 Niels Basjes
+# Copyright (C) 2013-2024 Niels Basjes
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -110,7 +110,7 @@ cat <<End-of-message
 # =============================================
 #
 # Yet Another UserAgent Analyzer
-# Copyright (C) 2013-2023 Niels Basjes
+# Copyright (C) 2013-2024 Niels Basjes
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.