+++ CHANGELOG.md
@@ -1171,7 +1171,7 @@ v0.1
 License
 =======
     Yet Another UserAgent Analyzer
-    Copyright (C) 2013-2023 Niels Basjes
+    Copyright (C) 2013-2024 Niels Basjes
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.